create template for profile
