+++ frontend/src/components/Settings.jsx
@@ -1,0 +1,9 @@
+const ApplicantActivity = () => {
+    return (
+        <div className="bg-white p-6 rounded-xl shadow-sm">
+            <h2 className="text-2xl font-semibold mb-4">Activity</h2>
+        </div>
+    );
+};
+
+export default ApplicantActivity;
